accepting const arrays in string creation functions
The functions create*String() take arrays of strings to be
used as the source for the newly-created Scilab variables.
The content of the source array is not modified during the
construction of a Scilab variable, but merely copied. Hence
it is possible to accept const arrays, which will enable
creation of Scilab variables from constant sources.

Change-Id: I4435174d6e1b375322bc0ec768da6c85e71eda29