GED: Bug fix to OpenGED buttons(toolbar/menubar)-Initialize in the property selected by user
Currently the buttons always open the property of the last figure.
This is a big problem because only the objects in the last window are accessible,
so the GED only worked well when there was only one open figure.

This commit resolves all problems!

With this commit ...
If an object was selected in the figure,
the GED is initiated with the property of this object.
Otherwise, it starts with the ownership of the current figure (that user pushed the button).

Change-Id: I232cee5d61e53c1f7fe4dd78eaaf108d2310dfdd