Move scirenderer code into Scilab source code.
The goal is to simplify the maintenance of Scirenderer with Scilab.
For now, during the release process, a lot of version of Scirenderer have to be published.
By this commit, we move the maintenance in a common place.
Scirenderer is STILL maintained as an independant project/library
Push from Scilab to the Scirenderer repository will be regularly done.

The configure argument "--without-embedded-scirenderer" bring back the same old behavior

Change-Id: I97bb98f1efb25eca31e7cc79ee4b291fbc48d0c8