* Bug #7486 fixed - Linear_algebra: DGELSY update and closer to the upstream (ZGELSY as well)
Update DGSELY (minor diff) from 3.0 to 3.3.1.

Instead of commenting out the part of the code that we are not interesed in, add an input parameter that will be tested around those parts.

This commit does not modifiy anything for the user, but we can now send our modifications to LAPACK.

Change-Id: I9ed03714943fd7fb443224dcccbb165df031f298