Xcos tests: recompilation at every call and context use
Modifiying the scs_m (Clock block period) requires recompilation of the cpr (Info),
so it was not safe to use the same Info variable.
Also, used the context to update that value,
instead of looking for it in the scs_m through loops.

In the diagrams, removed the unnecessary Scope blocks and recalibrated the buffer sizes
to match the minimal number of steps (final time / highest clock period).
Also, cleaned the contexts from unused variable L.

Change-Id: Ieeadd9f9c00f2e42273a3b12fb99e2e67e93e133