* Bug #9789 fixed - Optimization: optim printing its termination status
To know the termination status of optim, users had to get the output flag or
set imp>=1 as input argument.

Now, optim prints its termination status even when "imp" is not given
(imp=0 by default).

No nonreg test because the one mentioned on the bugzilla requires "uncprb",
and the unitary already meet that situation.

Change-Id: I5f4f7bfda3faa892791c9bdce9157357c5bf47ae