Optimization: clarify the imp<0 case of optim
In optim's calling sequence, when the imp flag is < 0,
meaning ind=1 in the user's function,
the modifications on f and/or g are ignored.

It is a case destined to other usages, such as iteration counting, information display,...

The "gc" solver was not adapted to that case because it only took ind=1 iterations when imp=-1.

Change-Id: Ief6d29fcab4f23ef22723591be431df257dfcb68