Graphics tests: fix plot3d_demo
Adapted plot3d_demo to the new plot3d() instruction
(after https://codereview.scilab.org/#/c/14174/),
that is, simply modified the window title.

I did not attach a .dia.ref file because the test still fails
because of sub_ticks, which are a priori not related to the above commit.
Proof of this is that plot2d and plot2d_demo tests fail for the same reason,
whereas the above commit does not affect them.

Change-Id: I57a6e9e9d5805685e80edbc4058610b367ba3501