* Bug #9623 fixed - Typo fixed: tics -> ticks.
This commit only fixes the part of Scilab visible by the user.

The source code (C, CPP, Java) is not modified, so the sub_ticks field is still available.
It is just not mentioned in the help.
Obsoleting that field requires quite a big change and would create many BC issues.

Change-Id: I82819ee75b43ca27ffbc0c84dfaeb63e66e6d03b