Xcos MVC: add a default explorer view
Implement a Treeshow-like Java view. This also check that everything is
fine while implementing a view with a state in the Event Dispatch
Thread.

Synchronization on Model object creation / deletion is performed with
the thread-safe shared_ptr. Any access on Controller::getProperty / Controller::setProperty with invalid uid will return an error status that is passed through the caller.

Change-Id: I79b474aa81db5eb365e2b755e757ccf4bebb0a68