Xcos MVC: add a default explorer view
Implement a Treeshow-like Java view. This also check that everything is
fine while implementing a view with a state in the Event Dispatch
Thread.

Synchronization on Model object creation / deletion should be managed
and a reference have to be stored on the Java side for all used object.
Any access on Controller::getProperty / Controller::setProperty with
invalid uid will return an error status that is passed through the
caller.

As the C -> Java calls are slow, the view is installed on demand. A huge
diagram can be loaded in seconds and the view retrieve data and track
all changes reducing performance once installed.

Change-Id: I79b474aa81db5eb365e2b755e757ccf4bebb0a68