Xcos MVC: reference counter instead of shared_ptr
shared_ptr can not be used on the Java side, instead of using 2
reference counting scheme (for composition and for multi-view) the
previous shared_ptr usage has been moved to the reference counting.

The API just allow any view to referenceObject which increment the
deleteObject number of calls needed to free the object.

Change-Id: I0c7d509bd7944c9044371ea2ae1b49900413071b