Xcos MVC: reference counter instead of shared_ptr
shared_ptr can not be used on the Java side, instead of using 2
reference counting scheme (for composition and for multi-view) the
previous shared_ptr usage has been moved to the reference counting.

The API just allow any view to referenceObject which increment the
deleteObject number of calls needed to free the object.

This commit also fix memory leaks by :

 * Allowing a cloneObject without deep copy of children.
 * Implementing DiagramAdapter cloning the right way (cloning children
   only once).
 * Re-using properties accessor on copy-constructors

Change-Id: I0c7d509bd7944c9044371ea2ae1b49900413071b