Adds a dynamic palette to show the "Recently used blocks"
We already save any tree modification (enable/disable or when we add some diagram
as a palette) back to the palette.xml file. So, we just need to change the tree
and Xcos will do the job to store it!

The palette browser will start normally, then if the user drag or press ENTER
some block, we create a new node on the tree, storing the block there.

We have a constant to define the maximum number of blocks. In addition, we
do not add duplicates.

Change-Id: I913a4b716d3c76a2dcbdd715b70b5942296e870f