Xcos: reintroduce a private xcosDiagramToScilab
importXcosDiagram use this helper to decode files using the Java
implementation. The rationale behind : `importXcosDiagram` is a public
function that manage all supported file types whereas `xcosDiagramToScilab`
is an helper used to call the Java implemented decoders.

NOTE: This commit also slightly improve the decoding of XCOS and ZCOS
diagrams.

Change-Id: I3f7329194e728e60bce8b47033b28b77969a9002