Xcos: put ANNOTATION at the end on sortAndFill
On saved files, the objects are sorted per kind. It is much more easier to
get the blocks first while debugging content. As the sorting is only
performed on the enum value, moving it will perform the job.

On the Xcos GUI and as the ANNOTATION will be decoded after the block, they
will appears "on top" of the BLOCK and LINK objects which is expected.

Change-Id: I966d4229e79f44021c416334c0105b30f894a7cb